Fix tests dependencies
