+++ res/mock_engine/rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "1.64.0"
+components = ["rustfmt", "clippy"]
+targets = ["wasm32-unknown-unknown"]
