+++ res/mock_eth_connector/rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "1.64.0"
+components = ["rustfmt", "clippy"]
+targets = ["wasm32-unknown-unknown"]
