+++ workspace/tests/utils.rs
@@ -6,12 +6,12 @@ const AURORA_LOCAL_CHAIN_ID: u64 = 1313161556;
 const AURORA_ACCOUNT_ID: &str = "aurora.test.near";
 const OWNER_ACCOUNT_ID: &str = "owner.test.near";
 const PROVER_ACCOUNT_ID: &str = "prover.test.near";
-#[allow(dead_code)]
-const EVM_CUSTODIAN_ADDRESS: &str = "096DE9C2B8A5B8c22cEe3289B101f6960d68E51E";
-const WASM_BIN_FILE_PATH: &str = "../../bin/mock_engine.wasm";
+const WASM_BIN_FILE_PATH: &str = "../bin/mock_engine.wasm";
 
 pub async fn init_and_deploy_contract() -> anyhow::Result<EvmContract> {
-    let worker = workspaces::sandbox().await?;
+    let worker = workspaces::sandbox()
+        .await
+        .map_err(|err| anyhow::anyhow!("Failed init sandbox: {:?}", err))?;
     let sk = SecretKey::from_random(KeyType::ED25519);
     let evm_account = worker
         .create_tla(AccountId::from_str(AURORA_ACCOUNT_ID)?, sk)
@@ -22,7 +22,7 @@ pub async fn init_and_deploy_contract() -> anyhow::Result<EvmContract> {
         owner_id: AccountId::from_str(OWNER_ACCOUNT_ID)?,
         prover_id: AccountId::from_str(PROVER_ACCOUNT_ID)?,
     };
-    let wasm = std::fs::read(WASM_BIN_FILE_PATH)?;
+    let wasm = std::fs::read(WASM_BIN_FILE_PATH).expect("failed read wasm file");
     // create contract
     let contract = EvmContract::deploy_and_init(evm_account, init_config, wasm).await?;
